ADB: Add adb backup for TWRP.
Functionality for client side to backup
tar and image streams over adbd to the client under backup.ab.

Using adb backup on the client side you can backup the partitions
TWRP knows about.

On the client side you can do the following:
adb backup -f <filename> --twrp <options> where options are
--compress: compress data
system: backup system
cache: backup cache
data: backup data
boot: backup boot
etc for each partition.

You can string multiple options,
i.e. adb backup -f <filename> --twrp --compress cache system data

adb backup in TWRP will take any option corresponding
to TWRP fstab partitions, e.g. efs boot as well.

If you do not specify the filename with the -f option,
adb will backup your data to a filename backup.ab on the client.
You can then rename the file and encrypt it with desktop tools.

If you don't want to use command line arguments:
adb backup --twrp

will bring up the gui and allow you to choose partitions
from the backup page.

To restore the backup use the following convention:
adb restore <filename>

Structures are used to store metadata in binary inside
of the file itself. If the metadata structure is modified,
update the adb version so that it will invalidate older
backups and not cause issues on restore. When restoring,
we currently do not support picking specific partitions.
It's all or nothing.

+++ progresstracking.hpp
@@ -45,7 +45,7 @@ private:
 	unsigned long long current_size;                   // Size of the current partition's already backed up data
 	unsigned long long current_count;                  // Count of files that have already been backed up for the current partition
 
-	unsigned long long previous_partitions_size;       // Total data already backed up from previous partitions (for the progress bar)
+	uint64_t previous_partitions_size;                 // Total data already backed up from previous partitions (for the progress bar)
 
 	bool display_file_count;                           // Inidicates if we will display the file count text
 	timespec last_update;                              // Tracks last update of the displayed progress (frequent updates tax the CPU and slow us down)
